--- src/components/objects/constants.js
@@ -1,2 +1,0 @@
-export const grapeProtocol = 'cg://'
-export const grapeProtocolRegExp = new RegExp(`^${grapeProtocol}`)
